Move additional Mod classes outside mod.cs
